renamed plugins module to 'imagingbook_calibrate_plugins'
