--- imagingbook-calibrate-plugins/src/main/java/module-info.java
@@ -1,10 +1,0 @@
-module imagingbook.calibrate.plugins {
-    exports Calibration_Plugins_1;
-    exports Calibration_Plugins_2;
-
-    requires ij;
-    requires java.desktop;
-    requires imagingbook.calibrate.lib;
-    requires imagingbook.common;
-    requires imagingbook.core;
-}
