+++ imagingbook_calibrate_plugins/.gitignore
