+++ imagingbook_calibrate_plugins/ImageJ/.gitignore
