+++ imagingbook_calibrate_plugins/ImageJ/plugins/plugins.config
@@ -1,0 +1,10 @@
+# plugins.config file for imagingbook_calibrate_plugins (automatically generated)
+# number of plugins: 7
+# date: 2023-01-09 18:08:03
+Plugins>Calibration Plugins 1, "Do Calibration", Calibration_Plugins_1.Do_Calibration
+Plugins>Calibration Plugins 1, "Open Image Stack", Calibration_Plugins_1.Open_Image_Stack
+Plugins>Calibration Plugins 1, "Validate EasyCalib Data", Calibration_Plugins_1.Validate_EasyCalib_Data
+Plugins>Calibration Plugins 2, "Draw 3D Axes Demo", Calibration_Plugins_2.Draw_3D_Axes_Demo
+Plugins>Calibration Plugins 2, "Rectify Camera Demo", Calibration_Plugins_2.Rectify_Camera_Demo
+Plugins>Calibration Plugins 2, "Replace Camera Demo", Calibration_Plugins_2.Replace_Camera_Demo
+Plugins>Calibration Plugins 2, "View Interpolation Demo", Calibration_Plugins_2.View_Interpolation_Demo
