+++ imagingbook_calibrate_plugins/pom.xml
@@ -1,0 +1,65 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	
+	<!-- Note: Name contains underscores such that the output JAR 
+	is compatible with the standard naming scheme for ImageJ plugins 
+	(see https://imagej.nih.gov/ij/plugins/jar-demo.html). -->
+	<artifactId>imagingbook_calibrate_plugins</artifactId>
+	<packaging>jar</packaging>
+	
+	<parent>
+		<groupId>com.imagingbook</groupId>
+		<artifactId>imagingbook-calibrate</artifactId>
+		<version>7.0.1-SNAPSHOT</version>
+	</parent>
+	
+	<properties>
+		<imagingbook.rootdir>${directory-up-2}</imagingbook.rootdir>
+		<maven.javadoc.skip>false</maven.javadoc.skip>
+	</properties>
+
+	<dependencies>
+	
+		<dependency>
+			<groupId>com.imagingbook</groupId>
+			<artifactId>imagingbook-calibrate-lib</artifactId>
+			<version>7.0.1-SNAPSHOT</version>
+		</dependency>
+
+		<dependency>
+			<groupId>com.imagingbook</groupId>
+			<artifactId>imagingbook-testing</artifactId>
+			<version>7.0.1-SNAPSHOT</version>
+			<scope>test</scope>
+		</dependency>
+
+		<dependency>
+			<groupId>junit</groupId>
+			<artifactId>junit</artifactId>
+			<version>${junit.version}</version>
+			<scope>test</scope>
+		</dependency>
+
+	</dependencies>
+	
+	<build>
+   		<!-- needed for updating Eclipse settings (see also ij-plugins-build profile) -->
+    	<outputDirectory>ImageJ/plugins</outputDirectory>
+    	
+    	<plugins>
+			<!-- do not deploy this artifact -->
+			<plugin>
+			  <groupId>org.apache.maven.plugins</groupId>
+			  <artifactId>maven-deploy-plugin</artifactId>
+			  <version>3.0.0-M2</version>
+			  <configuration>
+			    <skip>true</skip>
+			  </configuration>
+			</plugin>
+		</plugins>
+    </build>
+
+</project>
