+++ imagingbook_calibrate_plugins/src/main/java/Calibration_Plugins_1/package-info.java
@@ -1,0 +1,11 @@
+/*******************************************************************************
+ * Permission to use and distribute this software is granted under the BSD 2-Clause
+ * "Simplified" License (see http://opensource.org/licenses/BSD-2-Clause).
+ * Copyright (c) 2016-2023 Wilhelm Burger. All rights reserved.
+ * Visit https://imagingbook.com for additional details.
+ ******************************************************************************/
+
+@JavaDocBaseUrl("https://imagingbook.github.io/imagingbook-calibrate/javadoc/imagingbook_calibrate_plugins")
+package Calibration_Plugins_1;
+
+import imagingbook.core.jdoc.JavaDocBaseUrl;
