+++ imagingbook_calibrate_plugins/src/main/java/module-info.java
@@ -1,0 +1,10 @@
+module imagingbook_calibrate_plugins {
+    exports Calibration_Plugins_1;
+    exports Calibration_Plugins_2;
+
+    requires ij;
+    requires java.desktop;
+    requires imagingbook.calibrate.lib;
+    requires imagingbook.common;
+    requires imagingbook.core;
+}
