+++ imagingbook_calibrate_plugins/src/test/java/Calibration_Plugins_1/PackagePluginsTest.java
@@ -1,0 +1,28 @@
+/*******************************************************************************
+ * This software is provided as a supplement to the authors' textbooks on digital
+ * image processing published by Springer-Verlag in various languages and editions.
+ * Permission to use and distribute this software is granted under the BSD 2-Clause
+ * "Simplified" License (see http://opensource.org/licenses/BSD-2-Clause).
+ * Copyright (c) 2006-2023 Wilhelm Burger, Mark J. Burge. All rights reserved.
+ * Visit https://imagingbook.com for additional details.
+ ******************************************************************************/
+package Calibration_Plugins_1;
+
+import imagingbook.testutils.PluginTestUtils;
+import org.junit.Test;
+
+public class PackagePluginsTest {
+
+    private static Class<?>[] PLUGINS = {
+            Do_Calibration.class,
+            Open_Image_Stack.class
+    };
+
+    @Test
+    public void test1() throws Exception {
+        for (Class<?> clazz : PLUGINS) {
+            PluginTestUtils.CheckJavaDocHelp(clazz);
+        }
+    }
+
+}
